requirements.txt
spotipy
pandas
plotly
dash
python-dotenv
